+++ R/LearnerRegrXgboost.R
@@ -49,6 +49,14 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
+      p_nrounds = p_int(1L,
+        tags = c("train", "hotstart", "internal_tuning"),
+        aggr = crate(function(x) as.integer(ceiling(mean(unlist(x)))), .parent = topenv()),
+        in_tune_fn = crate(function(domain, param_vals) {
+          assert_true(!is.null(param_vals$early_stopping), .var.name = "early stopping rounds is set")
+          assert_integerish(domain$upper, len = 1L, any.missing = FALSE) }, .parent = topenv()),
+        disable_in_tune = list(early_stopping_rounds = NULL)
+      )
       ps = ps(
         alpha                       = p_dbl(0, default = 0, tags = "train"),
         approxcontrib               = p_lgl(default = FALSE, tags = "predict"),
@@ -80,7 +88,7 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
         missing                     = p_dbl(default = NA, tags = c("train", "predict"), special_vals = list(NA, NA_real_, NULL)),
         monotone_constraints        = p_uty(default = 0, tags = c("train", "control"), custom_check = crate(function(x) { checkmate::check_integerish(x, lower = -1, upper = 1, any.missing = FALSE) })), # nolint
         normalize_type              = p_fct(c("tree", "forest"), default = "tree", tags = "train", depends = quote(booster == "dart")),
-        nrounds                     = p_int(1L, tags = c("train", "hotstart")),
+        nrounds                     = p_nrounds,
         nthread                     = p_int(1L, default = 1L, tags = c("train", "threads")),
         ntreelimit                  = p_int(1, default = NULL, special_vals = list(NULL), tags = "predict"),
         num_parallel_tree           = p_int(1L, default = 1L, tags = "train"),
@@ -122,7 +130,7 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
         id = "regr.xgboost",
         param_set = ps,
         feature_types = c("logical", "integer", "numeric"),
-        properties = c("weights", "missings", "importance", "hotstart_forward"),
+        properties = c("weights", "missings", "importance", "hotstart_forward", "internal_tuning", "validation"),
         packages = c("mlr3learners", "xgboost"),
         label = "Extreme Gradient Boosting",
         man = "mlr3learners::mlr_learners_regr.xgboost"
@@ -249,12 +257,6 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
     }
   )
 )
-
-#' @export
-disable_internal_tuning.LearnerRegrXgboost = function(learner, ids) {
-  if (length(ids)) learner$param_set$set_values(early_stopping_rounds = NULL)
-  invisible(learner)
-}
 
 #' @export
 default_values.LearnerRegrXgboost = function(x, search_space, task, ...) { # nolint