+++ tests/testthat/test_regr_xgboost.R
@@ -57,7 +57,7 @@ test_that("validation and inner tuning", {
   expect_named(learner$model$evaluation_log, c("iter", "test_rmse"))
   expect_list(learner$internal_valid_scores, types = "numeric")
   expect_equal(names(learner$internal_valid_scores), "rmse")
-  expect_equal(learner$internal_valid_scores$logloss, learner$model$evaluation[get("iter") == 10, "test_rmse"][[1L]])
+  expect_equal(learner$internal_valid_scores$rmse, learner$model$evaluation[get("iter") == 10, "test_rmse"][[1L]])
 
   expect_list(learner$internal_tuned_values, types = "integerish")
   expect_equal(names(learner$internal_tuned_values), "nrounds")
@@ -75,4 +75,14 @@ test_that("validation and inner tuning", {
   expect_named(learner$model$evaluation_log, c("iter", "test_rmse"))
   expect_list(learner$internal_valid_scores, types = "numeric")
   expect_equal(names(learner$internal_valid_scores), "rmse")
+
+  learner = lrn("regr.xgboost",
+    nrounds = to_tune(upper = 1000, internal = TRUE),
+    validate = 0.2
+  )
+  s = learner$param_set$search_space()
+  expect_error(learner$param_set$convert_internal_search_space(s), "early stopping")
+  learner$param_set$set_values(early_stopping_rounds = 10)
+  learner$param_set$disable_internal_tuning("nrounds")
+  expect_equal(learner$param_set$values$early_stopping_rounds, NULL)
 })