+++ NEWS.md
@@ -1,5 +1,8 @@
 # mlr3learners (development version)
 
+* feat: `LearnerClassifXgboost` and `LearnerRegrXgboost` now support internal
+  tuning and validation. This now also works in conjunction with `mlr3pipelines`.
+
 # mlr3learners 0.5.8
 
 * Adaption to memory optimization in mlr3 0.17.1.